# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EternalHardcore Minecraft server API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ServerStatus {
  online: boolean;
  playerCount: number;
  maxPlayers: number;
  /** @nullable */
  latencyMs: number | null;
  motd: string;
  /**
   * Manual override: 'online' | 'offline' | null
   * @nullable
   */
  statusOverride: string | null;
}

export interface OnlinePlayer {
  username: string;
  /** @nullable */
  uuid: string | null;
  avatarUrl: string;
}

export interface PlayerStat {
  username: string;
  kills: number;
  deaths: number;
  playtimeMinutes: number;
  featured: boolean;
  avatarUrl: string;
}

export interface AddPlayerBody {
  username: string;
  kills: number;
  deaths: number;
  playtimeMinutes: number;
}

export interface StaffMember {
  username: string;
  role: string;
  avatarUrl: string;
}

export interface AddStaffBody {
  username: string;
  role: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  username: string;
  role: string;
}

export interface AdminInfo {
  username: string;
  role: string;
}

export interface ServerSettings {
  motd: string;
  /** @nullable */
  statusOverride: string | null;
  featuredPlayers: string[];
}

export interface UpdateServerSettingsBody {
  motd?: string;
  /** @nullable */
  statusOverride?: string | null;
  featuredPlayers?: string[];
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}