# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EternalHardcore Minecraft server API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns current server online status, player count, ping, and MOTD
 * @summary Get Minecraft server status
 */
export const GetServerStatusResponse = zod.object({
  online: zod.boolean(),
  playerCount: zod.number(),
  maxPlayers: zod.number(),
  latencyMs: zod.number().nullable(),
  motd: zod.string(),
  statusOverride: zod
    .string()
    .nullable()
    .describe("Manual override: 'online' | 'offline' | null"),
});

/**
 * Returns list of currently online players
 * @summary Get online players
 */
export const GetOnlinePlayersResponseItem = zod.object({
  username: zod.string(),
  uuid: zod.string().nullable(),
  avatarUrl: zod.string(),
});
export const GetOnlinePlayersResponse = zod.array(GetOnlinePlayersResponseItem);

/**
 * Returns player stats sorted by kills
 * @summary Get player stats leaderboard
 */
export const GetPlayerStatsResponseItem = zod.object({
  username: zod.string(),
  kills: zod.number(),
  deaths: zod.number(),
  playtimeMinutes: zod.number(),
  featured: zod.boolean(),
  avatarUrl: zod.string(),
});
export const GetPlayerStatsResponse = zod.array(GetPlayerStatsResponseItem);

/**
 * Authenticate as admin
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  username: zod.string(),
  role: zod.string(),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current admin session
 */
export const GetAdminMeResponse = zod.object({
  username: zod.string(),
  role: zod.string(),
});

/**
 * @summary Get server settings
 */
export const GetAdminSettingsResponse = zod.object({
  motd: zod.string(),
  statusOverride: zod.string().nullable(),
  featuredPlayers: zod.array(zod.string()),
});

/**
 * @summary Update server settings
 */
export const UpdateAdminSettingsBody = zod.object({
  motd: zod.string().optional(),
  statusOverride: zod.string().nullish(),
  featuredPlayers: zod.array(zod.string()).optional(),
});

export const UpdateAdminSettingsResponse = zod.object({
  motd: zod.string(),
  statusOverride: zod.string().nullable(),
  featuredPlayers: zod.array(zod.string()),
});

/**
 * @summary Get all players with stats (admin view)
 */
export const GetAdminPlayersResponseItem = zod.object({
  username: zod.string(),
  kills: zod.number(),
  deaths: zod.number(),
  playtimeMinutes: zod.number(),
  featured: zod.boolean(),
  avatarUrl: zod.string(),
});
export const GetAdminPlayersResponse = zod.array(GetAdminPlayersResponseItem);

/**
 * @summary Add a player to the leaderboard
 */
export const AddPlayerBody = zod.object({
  username: zod.string(),
  kills: zod.number(),
  deaths: zod.number(),
  playtimeMinutes: zod.number(),
});

export const AddPlayerResponse = zod.object({
  username: zod.string(),
  kills: zod.number(),
  deaths: zod.number(),
  playtimeMinutes: zod.number(),
  featured: zod.boolean(),
  avatarUrl: zod.string(),
});

/**
 * @summary Remove a player from the leaderboard
 */
export const DeletePlayerParams = zod.object({
  username: zod.coerce.string(),
});

export const DeletePlayerResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Toggle featured status for a player
 */
export const FeaturePlayerParams = zod.object({
  username: zod.coerce.string(),
});

export const FeaturePlayerResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get staff team list
 */
export const GetAdminStaffResponseItem = zod.object({
  username: zod.string(),
  role: zod.string(),
  avatarUrl: zod.string(),
});
export const GetAdminStaffResponse = zod.array(GetAdminStaffResponseItem);

/**
 * @summary Add or update a staff member
 */
export const AddStaffMemberBody = zod.object({
  username: zod.string(),
  role: zod.string(),
});

export const AddStaffMemberResponse = zod.object({
  username: zod.string(),
  role: zod.string(),
  avatarUrl: zod.string(),
});

/**
 * @summary Remove a staff member
 */
export const DeleteStaffMemberParams = zod.object({
  username: zod.coerce.string(),
});

export const DeleteStaffMemberResponse = zod.object({
  success: zod.boolean(),
});